# docs/swagger.yaml
info:
  contact: {}
paths:
  /cluster:
    delete:
      consumes:
      - application/json
      description: Delete a new cluster.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Delete a server cluster
      tags:
      - Cluster
    post:
      consumes:
      - application/json
      description: Create a new cluster.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Create a server cluster
      tags:
      - Cluster
  /example/helloworld:
    get:
      consumes:
      - application/json
      description: do ping
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: ping example
      tags:
      - example
  /information/GetTeams:
    get:
      consumes:
      - application/json
      description: Displays all team names attached to a username. Will use currently
        logged in user by default.
      parameters:
      - description: Username in DC1
        in: query
        name: username
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Get teams as designated by SNOW.
      tags:
      - Information
  /information/ProjectsByTeam:
    get:
      consumes:
      - application/json
      description: Get all projects by team requested.
      parameters:
      - description: Owning Team as defined in SNOW
        in: query
        name: owning_team
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Get projects by team ProjectsByTeam
      tags:
      - Information
  /information/ServersByProject:
    get:
      consumes:
      - application/json
      description: Displays all servers in a project.
      parameters:
      - description: Datacenter
        enum:
        - phx
        - sxb
        - iad
        in: query
        name: dc
        required: true
        type: string
      - description: Project Name
        in: query
        name: project
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Get servers by project
      tags:
      - Information
swagger: "2.0"